// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const CategoryStatus = {
  ACTIVE: 'ACTIVE',
  HIDDEN: 'HIDDEN'
} as const

export type CategoryStatus = (typeof CategoryStatus)[keyof typeof CategoryStatus]


export const ProductStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  ARCHIVED: 'ARCHIVED'
} as const

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus]


export const AttributeValueType = {
  NUMBER: 'NUMBER',
  TEXT: 'TEXT',
  ENUM: 'ENUM',
  BOOLEAN: 'BOOLEAN'
} as const

export type AttributeValueType = (typeof AttributeValueType)[keyof typeof AttributeValueType]


export const ImageKind = {
  PRODUCT: 'PRODUCT',
  LOGO: 'LOGO',
  DIAGRAM: 'DIAGRAM'
} as const

export type ImageKind = (typeof ImageKind)[keyof typeof ImageKind]


export const SourceKind = {
  MANUFACTURER: 'MANUFACTURER',
  AUTHORIZED_DISTRIBUTOR: 'AUTHORIZED_DISTRIBUTOR',
  RELIABLE: 'RELIABLE'
} as const

export type SourceKind = (typeof SourceKind)[keyof typeof SourceKind]


export const ProsConsKind = {
  PRO: 'PRO',
  CON: 'CON'
} as const

export type ProsConsKind = (typeof ProsConsKind)[keyof typeof ProsConsKind]